+++ pytsite/cache/_api.py
@@ -18,11 +18,27 @@ _dbg = _reg.get('cache.debug')
 def set_driver(driver: _AbstractDriver):
     """Register a cache driver
     """
+    global _current_driver
+
     if not isinstance(driver, _AbstractDriver):
         raise TypeError('Instance of {} expected, got {}'.format(_AbstractDriver, type(driver)))
 
-    global _current_driver
+    # When switching from one driver to another, it is important to move existing keys to the new storage,
+    keys_to_move = {}  # type: _Dict[str, list]
+    if _current_driver:
+        for pool in _pools.values():
+            keys_to_move[pool.uid] = []
+            for key in pool.keys():
+                keys_to_move[pool.uid].append((key, pool.get(key), pool.ttl(key)))  # Remember key and value
+                _pools[pool.uid].rm(key)  # Delete key via current driver
+
+    # Switch to the new driver
     _current_driver = driver
+
+    # Move keys from previous driver
+    for pool_uid, keys in keys_to_move.items():
+        for key, value, ttl in keys:
+            _pools[pool_uid].put(key, value, ttl)
 
 
 def get_driver() -> _AbstractDriver:
@@ -65,6 +81,12 @@ def get_pool(uid: str) -> _Pool:
 
     except KeyError:
         raise _error.PoolNotExist(uid)
+
+
+def get_pools() -> _Dict[str, _Pool]:
+    """Get all pools
+    """
+    return _pools.copy()
 
 
 def cleanup():